Add error handler class for AutoPkg. 	-- Restructure the project workspace folder structure
