+++ AutoPkgr/LGAutoPkgTask.m
@@ -18,6 +18,8 @@
 //
 
 #import "LGAutoPkgTask.h"
+#import "LGAutoPkgErrorHandler.h"
+
 #import "LGRecipes.h"
 #import "LGVersionComparator.h"
 #import "AHProxySettings.h"
@@ -57,6 +59,8 @@ typedef void (^AutoPkgReplyErrorBlock)(NSError *error);
 @property (copy, nonatomic, readwrite) NSMutableData *standardOutData;
 @property (copy, nonatomic, readwrite) NSString *standardOutString;
 @property (copy, nonatomic, readwrite) NSString *standardErrString;
+@property (strong, nonatomic) LGAutoPkgErrorHandler *errorHandler;
+
 @property (strong, nonatomic) LGVersioner *versioner;
 
 // Results objects
@@ -317,7 +321,7 @@ typedef void (^AutoPkgReplyErrorBlock)(NSError *error);
 
     if (!_error) {
         [self.taskLock lock];
-        self.error = [LGError errorWithTaskError:self.task verb:_verb];
+        self.error = [_errorHandler errorWithExitCode:self.task.terminationStatus];
         [self.taskLock unlock];
     }
 
@@ -344,7 +348,7 @@ typedef void (^AutoPkgReplyErrorBlock)(NSError *error);
 - (void)launchInBackground:(void (^)(NSError *))reply
 {
     LGAutoPkgTaskManager *bgQueue = [LGAutoPkgTaskManager new];
-    DLog(@"bgQueue: %@", bgQueue.name);
+    DevLog(@"bgQueue: %@", bgQueue.name);
     self.replyErrorBlock = reply;
     [bgQueue addOperation:self];
 }
@@ -397,14 +401,16 @@ typedef void (^AutoPkgReplyErrorBlock)(NSError *error);
 }
 
 #pragma mark - Task config helpers
-#pragma mark - Task config helpers
 - (void)configureFileHandles
 {
+    // Set up stdout
     NSPipe *standardOutput = [NSPipe pipe];
     self.task.standardOutput = standardOutput;
 
-    NSPipe *standardError = [NSPipe pipe];
-    self.task.standardError = standardError;
+    // Set up stderr
+    // The Error handler class creates a the pipe to process the stderr messages.
+    _errorHandler = [[LGAutoPkgErrorHandler alloc] initWithVerb:_verb];
+    self.task.standardError = _errorHandler.pipe;
 
     if (_verb == kLGAutoPkgRun || _verb == kLGAutoPkgRepoUpdate) {
         if (self.AUTOPKG_VERSION_0_4_0) {
@@ -424,43 +430,48 @@ typedef void (^AutoPkgReplyErrorBlock)(NSError *error);
 
             BOOL verbose = [[NSUserDefaults standardUserDefaults] boolForKey:@"verboseAutoPkgRun"];
             [[standardOutput fileHandleForReading] setReadabilityHandler:^(NSFileHandle *handle) {
-                NSString *message = [[NSString alloc]initWithData:[handle availableData] encoding:NSUTF8StringEncoding];
-                
-                [_versioner parseString:message];
-                if ([progressPredicate evaluateWithObject:message]) {
-                    NSString *fullMessage;
-                    if (_verb == kLGAutoPkgRepoUpdate) {
-                        fullMessage = [NSString stringWithFormat:@"Updating %@", [message lastPathComponent]];
-                    } else {
-                        int cntStr = (int)round(count) + 1;
-                        int totStr = (int)round(total);
-                        fullMessage = [[NSString stringWithFormat:@"(%d/%d) %@", cntStr, totStr, message] stringByTrimmingCharactersInSet:[NSCharacterSet newlineCharacterSet]];
+                NSData *data = handle.availableData;
+                if (data.length) {
+                    NSString *message = [[NSString alloc]initWithData:data encoding:NSUTF8StringEncoding];
+
+                    [_versioner parseString:message];
+                    if ([progressPredicate evaluateWithObject:message]) {
+                        NSString *fullMessage;
+                        if (_verb == kLGAutoPkgRepoUpdate) {
+                            fullMessage = [NSString stringWithFormat:@"Updating %@", [message lastPathComponent]];
+                        } else {
+                            int cntStr = (int)round(count) + 1;
+                            int totStr = (int)round(total);
+                            fullMessage = [[NSString stringWithFormat:@"(%d/%d) %@", cntStr, totStr, message] stringByTrimmingCharactersInSet:[NSCharacterSet newlineCharacterSet]];
+                        }
+
+                        double progress = ((count/total) * 100);
+
+                        LGAutoPkgTaskResponseObject *response = [[LGAutoPkgTaskResponseObject alloc] init];
+                        response.progressMessage = fullMessage;
+                        response.progress = progress;
+                        count++;
+
+                        [(NSObject *)_taskStatusDelegate performSelectorOnMainThread:@selector(didReceiveStatusUpdate:) withObject:response waitUntilDone:NO];
+
+                        // If verboseAutoPkgRun is not enabled, log the limited message here.
+                        if (!verbose) {
+                            NSLog(@"%@",message);
+                        }
                     }
-
-                    double progress = ((count/total) * 100);
-
-                    LGAutoPkgTaskResponseObject *response = [[LGAutoPkgTaskResponseObject alloc] init];
-                    response.progressMessage = fullMessage;
-                    response.progress = progress;
-                    count++;
-
-                    [(NSObject *)_taskStatusDelegate performSelectorOnMainThread:@selector(didReceiveStatusUpdate:) withObject:response waitUntilDone:NO];
-
-                    // If verboseAutoPkgRun is not enabled, log the limited message here.
-                    if (!verbose) {
+                    // If verboseAutoPkgRun is enabled, log everything generated by autopkg run -v.
+                    if (verbose) {
                         NSLog(@"%@",message);
                     }
-                }
-                // If verboseAutoPkgRun is enabled, log everything generated by autopkg run -v.
-                if (verbose) {
-                    NSLog(@"%@",message);
                 }
             }];
         }
     } else {
+        // In order to prevent maxing out the stdout buffer collect the data progressively
+        // even thought the data returned is usually small.
         [[standardOutput fileHandleForReading] setReadabilityHandler:^(NSFileHandle *fh) {
             NSData *data = [fh availableData];
-            if ([data length]) {
+            if (data.length) {
                 if (!_standardOutData) {
                     _standardOutData = [[NSMutableData alloc] init ];
                 }
@@ -525,18 +536,22 @@ typedef void (^AutoPkgReplyErrorBlock)(NSError *error);
 }
 
 #pragma mark - Output / Results
+- (int)exitCode
+{
+    if (!self.task.isRunning) {
+        return self.task.terminationStatus;
+    }
+
+    return NSTaskTerminationReasonUncaughtSignal;
+}
+
 - (NSString *)standardErrString
 {
     [self.taskLock lock];
     if (!_standardErrString && !self.task.isRunning) {
-        NSData *data;
-        if ([self.task.standardError isKindOfClass:[NSPipe class]]) {
-            data = [[self.task.standardError fileHandleForReading] readDataToEndOfFile];
-            if (data) {
-                _standardErrString = [[NSString alloc] initWithData:data encoding:NSASCIIStringEncoding];
-            }
-        }
+        _standardErrString = _errorHandler.errorString;
     }
+
     [self.taskLock unlock];
     return _standardErrString;
 }
